# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: string;
  name: string;
  genre: string;
  status: string;
  progress: number;
  description: string;
  updatedAt: string;
  isExample: boolean;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  genre: string;
  description: string;
}

export interface HealthMetric {
  label: string;
  value: number;
  color: string;
}

export interface DashboardStats {
  characters: number;
  chapters: number;
  scenes: number;
  decisions: number;
  endings: number;
  issues: number;
}

export interface Activity {
  id: string;
  type: string;
  title: string;
  detail: string;
  timestamp: string;
}

export interface Dashboard {
  project: Project;
  stats: DashboardStats;
  health: HealthMetric[];
  activity: Activity[];
}

export interface Character {
  id: string;
  name: string;
  role: string;
  initials: string;
  description: string;
  arc: string;
  tags: string[];
  updatedAt: string;
}

export interface CharacterInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  description: string;
  arc: string;
  tags?: string[];
}

export interface Chapter {
  id: string;
  number: number;
  title: string;
  status: string;
  sceneCount: number;
  summary: string;
}

export interface Scene {
  id: string;
  chapterId: string;
  number: number;
  title: string;
  location: string;
  kind: string;
  status: string;
  updatedAt: string;
  preview: string;
}

export type GetDashboardParams = {
projectId?: string;
};

export type ListCharactersParams = {
projectId?: string;
};

export type CreateCharacterParams = {
projectId?: string;
};

export type ListChaptersParams = {
projectId?: string;
};

export type ListScenesParams = {
projectId?: string;
};