# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: workspace
    description: Narrative Forge workspace operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [workspace]
      summary: Get the narrative workspace dashboard
      parameters:
        - in: query
          name: projectId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /projects:
    get:
      operationId: listProjects
      tags: [workspace]
      summary: List narrative projects
      responses:
        "200":
          description: Narrative projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [workspace]
      summary: Create a narrative project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created narrative project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /characters:
    get:
      operationId: listCharacters
      tags: [workspace]
      summary: List project characters
      parameters:
        - in: query
          name: projectId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Character database
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
    post:
      operationId: createCharacter
      tags: [workspace]
      summary: Create a character
      parameters:
        - in: query
          name: projectId
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
  /chapters:
    get:
      operationId: listChapters
      tags: [workspace]
      summary: List story chapters
      parameters:
        - in: query
          name: projectId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Story chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
  /scenes:
    get:
      operationId: listScenes
      tags: [workspace]
      summary: List recent story scenes
      parameters:
        - in: query
          name: projectId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Story scenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scene"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Project:
      type: object
      required: [id, name, genre, status, progress, description, updatedAt, isExample]
      properties:
        id: { type: string }
        name: { type: string }
        genre: { type: string }
        status: { type: string }
        progress: { type: number }
        description: { type: string }
        updatedAt: { type: string }
        isExample: { type: boolean }
    ProjectInput:
      type: object
      required: [name, genre, description]
      properties:
        name: { type: string, minLength: 1 }
        genre: { type: string, minLength: 1 }
        description: { type: string }
    HealthMetric:
      type: object
      required: [label, value, color]
      properties:
        label: { type: string }
        value: { type: number }
        color: { type: string }
    DashboardStats:
      type: object
      required: [characters, chapters, scenes, decisions, endings, issues]
      properties:
        characters: { type: number }
        chapters: { type: number }
        scenes: { type: number }
        decisions: { type: number }
        endings: { type: number }
        issues: { type: number }
    Activity:
      type: object
      required: [id, type, title, detail, timestamp]
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string }
        detail: { type: string }
        timestamp: { type: string }
    Dashboard:
      type: object
      required: [project, stats, health, activity]
      properties:
        project:
          $ref: "#/components/schemas/Project"
        stats:
          $ref: "#/components/schemas/DashboardStats"
        health:
          type: array
          items:
            $ref: "#/components/schemas/HealthMetric"
        activity:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
    Character:
      type: object
      required: [id, name, role, initials, description, arc, tags, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        initials: { type: string }
        description: { type: string }
        arc: { type: string }
        tags:
          type: array
          items: { type: string }
        updatedAt: { type: string }
    CharacterInput:
      type: object
      required: [name, role, description, arc]
      properties:
        name: { type: string, minLength: 1 }
        role: { type: string, minLength: 1 }
        description: { type: string }
        arc: { type: string }
        tags:
          type: array
          items: { type: string }
    Chapter:
      type: object
      required: [id, number, title, status, sceneCount, summary]
      properties:
        id: { type: string }
        number: { type: number }
        title: { type: string }
        status: { type: string }
        sceneCount: { type: number }
        summary: { type: string }
    Scene:
      type: object
      required: [id, chapterId, number, title, location, kind, status, updatedAt, preview]
      properties:
        id: { type: string }
        chapterId: { type: string }
        number: { type: number }
        title: { type: string }
        location: { type: string }
        kind: { type: string }
        status: { type: string }
        updatedAt: { type: string }
        preview: { type: string }